+++ src/main/java/chess/controller/ChessGameController.java
@@ -3,9 +3,11 @@ package chess.controller;
 import static spark.Spark.get;
 import static spark.Spark.post;
 
-import chess.controller.dto.ChessGameResponse;
-import chess.controller.dto.PromotionRequest;
 import chess.controller.dto.StatusResponse;
+import chess.controller.dto.request.MoveRequest;
+import chess.controller.dto.request.PromotionRequest;
+import chess.controller.dto.response.ChessGameResponse;
+import chess.domain.Position;
 import chess.domain.PromotionPiece;
 import chess.service.ChessGameService;
 import com.google.gson.Gson;
@@ -29,6 +31,14 @@ public class ChessGameController {
             PromotionRequest promotionRequest = gson.fromJson(req.body(), PromotionRequest.class);
             PromotionPiece promotionPiece = PromotionPiece.createPromotionPiece(promotionRequest.getPromotionValue());
             chessGameService.promotion(promotionPiece);
+            return StatusResponse.SUCCESS;
+        });
+
+        post("/move", "application/json", (req, res) -> {
+            MoveRequest moveRequest = gson.fromJson(req.body(), MoveRequest.class);
+            Position source = moveRequest.toSourcePosition();
+            Position target = moveRequest.toTargetPosition();
+            chessGameService.move(source, target);
             return StatusResponse.SUCCESS;
         });
     }
