+++ src/main/java/chess/controller/dto/request/PromotionRequest.java
@@ -1,4 +1,4 @@
-package chess.controller.dto;
+package chess.controller.dto.request;
 
 public class PromotionRequest {
