+++ src/main/java/chess/controller/dto/response/PieceResponse.java
@@ -1,4 +1,4 @@
-package chess.controller.dto;
+package chess.controller.dto.response;
 
 import chess.domain.piece.Piece;
 import java.util.Locale;
